Proyecto/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

# =========================================================
# Project metadata
# =========================================================
[project]
name = "final-project-pipeline"
version = "0.1.0"
description = "End-to-end data engineering pipeline project template (GCP reference) using Python"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }

authors = [
  { name = "BSG Institute" }
]

keywords = [
  "data-engineering",
  "etl",
  "pipeline",
  "gcp",
  "bigquery",
  "gcs",
  "bronze-silver-gold"
]

classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Education",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent"
]

# =========================================================
# Core dependencies (runtime)
# =========================================================
dependencies = [
  "pandas>=2.2.0",
  "pyarrow>=16.0.0",
  "requests>=2.32.0",
  "python-dotenv>=1.0.1",
  "tenacity>=8.3.0",

  # GCP reference
  "google-cloud-storage>=2.18.0",
  "google-cloud-bigquery[pandas]>=3.25.0",

  # Structured logging (optional but recommended)
  "structlog>=24.1.0"
]

# =========================================================
# Optional dependencies (extras)
# =========================================================
[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-cov>=5.0.0",
  "ruff>=0.3.0",
  "black>=24.0.0",
  "mypy>=1.9.0",
  "ipykernel>=6.29.0"
]

# =========================================================
# setuptools configuration (src layout)
# =========================================================
[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

# (opcional) incluir archivos no-Python si los llegas a empaquetar
# [tool.setuptools.package-data]
# pipeline = ["config/*.yml", "sql/*.sql"]

# =========================================================
# Tooling configuration
# =========================================================
[tool.black]
line-length = 100
target-version = ["py310"]
exclude = '''
/(
    \.git
  | \.venv
  | build
  | dist
  | out
)/
'''

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
warn_unused_ignores = true

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q"
testpaths = ["tests"]

[tool.coverage.run]
branch = true
source = ["src"]

[tool.coverage.report]
show_missing = true
skip_empty = true
